+++ mocket/core/inject.py
@@ -33,10 +33,10 @@ def enable() -> None:
         mock_socketpair,
     )
     from mocket.core.ssl.context import MocketSSLContext, mock_wrap_socket
-    from mocket.urllib3 import (
+    from mocket.core.urllib3 import (
         mock_match_hostname as mock_urllib3_match_hostname,
     )
-    from mocket.urllib3 import (
+    from mocket.core.urllib3 import (
         mock_ssl_wrap_socket as mock_urllib3_ssl_wrap_socket,
     )
