+++ mocket/core/urllib3.py
